refactor(slider): Barrel file renaming
