+++ projects/igniteui-angular/src/lib/slider/public_api.ts
